api.py modularized
